claude fix 4

+++ download_models.py
@@ -1,29 +1,18 @@
-import torch
 from qwen_tts import Qwen3TTSModel
 import whisper_timestamped as whisper
 
-# ✅ FIX: Use `dtype` instead of `torch_dtype`.
-# The Qwen3 TTS library deprecated `torch_dtype` and silently ignores it,
-# meaning models were being cached as float32. Using `dtype=torch.float16`
-# ensures the cached weights match what the handler loads at runtime.
+# Download and cache model weights to disk.
+# Note: we do NOT convert dtype here — that happens at runtime via .half().cuda()
+# in handler.py, which reliably converts ALL sub-models regardless of kwarg support.
 
 print("--- ⬇️ Downloading Qwen VoiceDesign ---")
-Qwen3TTSModel.from_pretrained(
-    "Qwen/Qwen3-TTS-12Hz-1.7B-VoiceDesign",
-    dtype=torch.float16,
-)
+Qwen3TTSModel.from_pretrained("Qwen/Qwen3-TTS-12Hz-1.7B-VoiceDesign")
 
 print("--- ⬇️ Downloading Qwen CustomVoice ---")
-Qwen3TTSModel.from_pretrained(
-    "Qwen/Qwen3-TTS-12Hz-1.7B-CustomVoice",
-    dtype=torch.float16,
-)
+Qwen3TTSModel.from_pretrained("Qwen/Qwen3-TTS-12Hz-1.7B-CustomVoice")
 
 print("--- ⬇️ Downloading Qwen Base (Clone) ---")
-Qwen3TTSModel.from_pretrained(
-    "Qwen/Qwen3-TTS-12Hz-1.7B-Base",
-    dtype=torch.float16,
-)
+Qwen3TTSModel.from_pretrained("Qwen/Qwen3-TTS-12Hz-1.7B-Base")
 
 print("--- ⬇️ Downloading Whisper ---")
 whisper.load_model("base", device="cpu")
